requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
captum==0.8.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.5.1
grpcio==1.73.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
lightning-utilities==0.14.3
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pillow==11.2.1
propcache==0.3.2
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
PyYAML==6.0.2
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
torch==2.7.1
torchmetrics==1.7.2
tqdm==4.67.1
typing_extensions==4.14.0
Werkzeug==3.1.3
yarl==1.20.1
